requirements.txt
amqp==2.6.1
asgiref==3.2.10
Babel==2.8.0
billiard==3.6.3.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.2
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.0.8
django-admin-rangefilter==0.6.1
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-jazzmin==2.2.5
django-phone-field==1.8.1
django-phonenumber-field==4.0.0
flower==0.9.3
html5lib==1.1
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
oauthlib==3.1.0
Pillow==7.2.0
pycparser==2.20
Pyphen==0.9.5
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
tinycss2==1.0.2
tornado==5.1.1
urllib3==1.25.10
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
zipp==3.1.0
